requirements.txt
tensorflow
pandas
numpy
scikit-learn
streamlit
faiss-cpu
gradio
python-dotenv
requests